Extract 'FeedCacheSpy' into shared scope to remove duplication.
