+++ EssentailFeed/EssentailFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -123,47 +123,6 @@ class CacheFeedUseCaseTests: XCTestCase {
         XCTAssertEqual(receivedError as NSError?, expectedError, file: file, line: line)
     }
     
-    private class FeedStoreSpy: FeedStore {
-
-        typealias DeletionCompletion = (Error?) -> Void
-        typealias InsertionCompletion = (Error?) -> Void
-
-        enum ReceivedMessage : Equatable {
-            case deletCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        private(set) var receivedMessage  = [ReceivedMessage]()
-        
-        private var delectCompletions = [DeletionCompletion]()
-        private var insertionCompletions = [InsertionCompletion]()
-
-        func deletionCacheFeed(completion: @escaping DeletionCompletion){
-            delectCompletions.append(completion)
-            receivedMessage.append(.deletCachedFeed)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            delectCompletions[index](error)
-        }
-        
-        func completeDeletionSuccessfully(at index: Int = 0){
-            delectCompletions[index](nil)
-        }
-        
-        func insert(_ feed: [LocalFeedImage],timestamp: Date, completion: @escaping InsertionCompletion) {
-            insertionCompletions.append(completion)
-            receivedMessage.append(.insert(feed, timestamp))
-        }
-        
-        func completeInsetion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsetionSuccessfully(at index: Int = 0){
-            insertionCompletions[index](nil)
-        }
-    }
-    
     private func uniqueImage() -> FeedImage {
         return FeedImage(id: UUID(), description: "any desc", location: "any loc", url: anyURL())
     }
